# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payout API Documentation",
        "description": "API documentation for MFS and Bank Payout System",
        "contact": {
            "email": "support@faastpay.xyz"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://payoutsystem.faastpay.xyz"
        },
        {
            "url": "https://payoutsystem.faastpay.xyz"
        }
    ],
    "paths": {
...
        "/api/merchant/payouts": {
            "post": {
                "tags": [
                    "Bank Payouts"
                ],
                "summary": "Create a new Bank Payout Batch",
                "operationId": "160af7702b43c9be78c50a5c13d47eee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "merchant_id",
                                    "batch_id",
                                    "payouts"
                                ],
                                "properties": {
                                    "merchant_id": {
                                        "type": "string",
                                        "example": "faastpay"
                                    },
                                    "batch_id": {
                                        "type": "string",
                                        "example": "BANK-ABC123XYZ"
                                    },
                                    "payouts": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "referenceKey": {
                                                    "type": "string",
                                                    "example": "BANK-REF001"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 500
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "BDT"
                                                },
                                                "remarks": {
                                                    "type": "string",
                                                    "example": "Vendor payment"
                                                },
                                                "bankCode": {
                                                    "type": "string",
                                                    "example": "011"
                                                },
                                                "bankShortCode": {
                                                    "type": "string",
                                                    "example": "DBBL"
                                                },
                                                "benType": {
                                                    "type": "string",
                                                    "example": "individual"
                                                },
                                                "txnChannel": {
                                                    "type": "string",
                                                    "example": "BEFTN"
                                                },
                                                "beneficiaryAcc": {
                                                    "type": "string",
                                                    "example": "123456789"
                                                },
                                                "beneficiaryName": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "beneficiaryEmail": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "routingNumber": {
                                                    "type": "string",
                                                    "example": "123456789"
                                                },
                                                "txnChannelCode": {
                                                    "type": "string",
                                                    "example": "beftn"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payouts created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payouts created for batch: BANK-ABC123XYZ"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Insufficient balance"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/merchant/payouts/batch/{batch_id}": {
            "get": {
                "tags": [
                    "Bank Payouts"
                ],
                "summary": "Fetch all payouts under a given batch ID",
                "operationId": "b7b795f2ff336a2408ffa13f36168ead",
                "parameters": [
                    {
                        "name": "batch_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Batch details fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "batch_id": {
                                            "type": "string",
                                            "example": "BANK-ABC123XYZ"
                                        },
                                        "total_count": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "total_amount": {
                                            "type": "number",
                                            "example": 1000
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Batch not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/merchant/payouts/reference/{referenceKey}": {
            "get": {
                "tags": [
                    "Bank Payouts"
                ],
                "summary": "Fetch a single payout using the reference key",
                "operationId": "2d27b1045f98b0f730e4e0efbb5a78fc",
                "parameters": [
                    {
                        "name": "referenceKey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payout found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "referenceKey": {
                                                    "type": "string",
                                                    "example": "BANK-REF001"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 500
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "BDT"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "Pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payout not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}